refactor(tec1g): decompose runtime into config/state/matrix/storage/lifecycle modules (#227)
- Move config normalization to runtime-config.ts
- Move runtime state/type initialization to runtime-state.ts
- Move matrix staging/idle commit logic to runtime-matrix.ts
- Move SD image load+persistence wiring to runtime-storage.ts
- Move reset/silence lifecycle logic to runtime-lifecycle.ts
- Keep runtime.ts as orchestration and re-export normalizeTec1gConfig/Tec1gState

Fixes #227

Made-with: Cursor
