+++ src/platforms/tec1g/runtime-config.ts
@@ -1,0 +1,83 @@
+/**
+ * @file TEC-1G runtime configuration normalization.
+ */
+
+import type { Tec1gPlatformConfig, Tec1gPlatformConfigNormalized } from '../types';
+import { normalizeSimpleRegions } from '../simple/runtime';
+import {
+  TEC1G_ADDR_MAX,
+  TEC1G_APP_START_DEFAULT,
+  TEC1G_ENTRY_DEFAULT,
+  TEC1G_RAM_END,
+  TEC1G_RAM_START,
+  TEC1G_ROM0_END,
+  TEC1G_ROM0_START,
+  TEC1G_ROM1_END,
+  TEC1G_ROM1_START,
+} from './constants';
+
+/**
+ * Normalizes TEC-1G configuration with defaults and bounds.
+ */
+export function normalizeTec1gConfig(cfg?: Tec1gPlatformConfig): Tec1gPlatformConfigNormalized {
+  const config: Tec1gPlatformConfig = cfg ?? {};
+  const regions = normalizeSimpleRegions(config.regions, [
+    { start: TEC1G_ROM0_START, end: TEC1G_ROM0_END, kind: 'rom' },
+    { start: TEC1G_RAM_START, end: TEC1G_RAM_END, kind: 'ram' },
+    { start: TEC1G_ROM1_START, end: TEC1G_ROM1_END, kind: 'rom' },
+  ]);
+  const romRanges = regions
+    .filter((region) => region.kind === 'rom' || region.readOnly === true)
+    .map((region) => ({ start: region.start, end: region.end }));
+  const appStart =
+    Number.isFinite(config.appStart) && config.appStart !== undefined
+      ? config.appStart
+      : TEC1G_APP_START_DEFAULT;
+  const entry =
+    Number.isFinite(config.entry) && config.entry !== undefined ? config.entry : TEC1G_ENTRY_DEFAULT;
+  const romHex = typeof config.romHex === 'string' && config.romHex !== '' ? config.romHex : undefined;
+  const ramInitHex =
+    typeof config.ramInitHex === 'string' && config.ramInitHex !== '' ? config.ramInitHex : undefined;
+  const cartridgeHex =
+    typeof config.cartridgeHex === 'string' && config.cartridgeHex !== ''
+      ? config.cartridgeHex
+      : undefined;
+  const updateMs =
+    Number.isFinite(config.updateMs) && config.updateMs !== undefined ? config.updateMs : 16;
+  const yieldMs = Number.isFinite(config.yieldMs) && config.yieldMs !== undefined ? config.yieldMs : 0;
+  const extraListings = Array.isArray(config.extraListings)
+    ? config.extraListings
+        .map((entry) => (typeof entry === 'string' ? entry.trim() : ''))
+        .filter((entry) => entry !== '')
+    : undefined;
+  const gimpSignal = config.gimpSignal === true;
+  const expansionBankHi = config.expansionBankHi === true;
+  const matrixMode = config.matrixMode === true;
+  const protectOnReset = config.protectOnReset === true;
+  const rtcEnabled = config.rtcEnabled === true;
+  const sdEnabled = config.sdEnabled === true;
+  const sdHighCapacity = config.sdHighCapacity !== false;
+  const sdImagePath =
+    typeof config.sdImagePath === 'string' && config.sdImagePath !== '' ? config.sdImagePath : undefined;
+  return {
+    regions,
+    romRanges,
+    appStart: Math.max(0, Math.min(TEC1G_ADDR_MAX, appStart)),
+    entry: Math.max(0, Math.min(TEC1G_ADDR_MAX, entry)),
+    ...(romHex !== undefined ? { romHex } : {}),
+    ...(ramInitHex !== undefined ? { ramInitHex } : {}),
+    updateMs: Math.max(0, updateMs),
+    yieldMs: Math.max(0, yieldMs),
+    gimpSignal,
+    expansionBankHi,
+    matrixMode,
+    protectOnReset,
+    rtcEnabled,
+    sdEnabled,
+    sdHighCapacity,
+    ...(sdImagePath !== undefined ? { sdImagePath } : {}),
+    ...(cartridgeHex !== undefined ? { cartridgeHex } : {}),
+    ...(extraListings ? { extraListings } : {}),
+    ...(cfg?.uiVisibility ? { uiVisibility: cfg.uiVisibility } : {}),
+  };
+}
